product-catalog-service/pyproject.toml
[project]
name = "product-catalog-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
